backend/requirements.txt
fastapi
uvicorn[standard]
openai-whisper
TTS
python-multipart
pydub
spacy
SQLAlchemy
resemblyzer